starter-data-jpa

+++ src/main/java/org/example/spring/dto/CompanyReadDto.java
@@ -1,4 +1,4 @@
 package org.example.spring.dto;
 
-public record CompanyReadDto(Integer id) {
+public record CompanyReadDto(Long id) {
 }
