+++ src/main/java/org/example/spring/service/UserService.java
@@ -11,6 +11,6 @@ import org.springframework.stereotype.Service;
 public class UserService {
     private final UserRepository userRepository;
 
-    private final CrudRepository<Integer, Company> companyRepository ;
+    private final CrudRepository<Long, Company> companyRepository ;
 
 }
